add ltdiff to page as output on analysis

+++ domain/src/main/java/sogyo/wesley/ipsportdata/domain/AnalysisFactory.java
@@ -3,7 +3,7 @@ package sogyo.wesley.ipsportdata.domain;
 public class AnalysisFactory implements IFactory {
 
     @Override
-    public IAnalysor createNewAnalysis(int power, float lactate_one, float lactate_two) {
+    public IAnalysor createNewAnalysis(int power, double lactate_one, double lactate_two) {
         return new Analysor(power, lactate_one, lactate_two);
     }
 
@@ -13,3 +13,4 @@ public class AnalysisFactory implements IFactory {
     }
 
   }
+
