+++ domain/src/main/java/sogyo/wesley/ipsportdata/domain/IFactory.java
@@ -1,6 +1,6 @@
 package sogyo.wesley.ipsportdata.domain;
 
 public interface IFactory {
-    IAnalysor createNewAnalysis(int power, float lactate_one, float lactate_two);
+    IAnalysor createNewAnalysis(int power, double lactate_one, double lactate_two);
     IAnalysor createNewAnalysis(String name);
 }
